exe 2
This second challenge will help you to re-cap your CSS text formatting skills, as well as start using margins and padding. Use shakespeareSonnet.html to get started, and the image above is what your page should look like.


Hints:

The thumbnails are not very representative, but this challenge requires you to create a fixed-width layout that is centered in the browser window.
To get the wrap division to stay centered, you'll need to set the left and right margins to a_t_ (you've got to fill in the blanks).